Added changing is_verified and changed_date

+++ account/views.py
@@ -289,27 +289,30 @@ class AccountSendVerifyRequest(APIView):
             return response
         else:
             response = Response({'detail':'you cannot send a verify request'},
-                                status=status.HTTP_200_OK)
+                                status=status.HTTP_400_BAD_REQUEST)
             return response
         
 class AccountChangeVerify(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request, id, format=None):
-        if request.user.is_moderator:
-            verify_status = request.data.get('verify')
-            verify_request = VerifiedAccount.objects.get(id=id)
+        serializer = AccountChangeVerifySerializer(data=request.data)
+        if serializer.is_valid(raise_exception=True):
+            if request.user.is_moderator:
+                verify_status = serializer.validated_data.get('is_verified')
+                verify_request = VerifiedAccount.objects.get(id=id)
 
-            if verify_status == True:
-                verify_request.is_verified = True
-                verify_request.save()
+                if verify_status == True:
+                    verify_request.is_verified = True
+                    verify_request.changed_date = datetime.datetime.now
+                    verify_request.save()
+                else:
+                    verify_request.delete()
+
+                response = Response({'detail':'verify account has been changed'},
+                                    status=status.HTTP_200_OK)
+                return response
             else:
-                verify_request.delete()
-
-            response = Response({'detail':'verify account has been changed'},
-                                status=status.HTTP_200_OK)
-            return response
-        else:
-            response = Response({'detail':'you cannot change a verify account'},
-                                status=status.HTTP_200_OK)
-            return response
+                response = Response({'detail':'you cannot change a verify account'},
+                                    status=status.HTTP_400_BAD_REQUEST)
+                return response
